## Description

This pr makes the query faster when a problem report is generated.

- [Jira](https://edunext.atlassian.net/browse/PS2021-1174?atlOrigin=eyJpIjoiYzI4NDU5ODlkNzY2NDEyNDhjNzlmZGQzODZkNzA1MTQiLCJwIjoiaiJ9)
- [Juniper version](https://github.com/eduNEXT/edunext-platform/pull/445)
- [Documentation](https://docs.google.com/document/d/1wCxCeqojOizVbF0JJCwfCv75bR1XTnjTdVkzZv5mQ6U/edit)

## How to test

To test go to the lms to a course and in the instructor and data download tab select "create a report of problem ..."
Example: https://bux-staging-bracu-ac-bd.le.edunextstage.net/courses/course-v1:buX+ARC101+2021_Spring/instructor#view-data_download
![Screenshot from 2021-12-08 18-39-13](https://user-images.githubusercontent.com/35668326/145302858-1c251626-e89b-435a-960b-30e979850b5d.png)

## Result of my tests

- https://bux-staging-bracu-ac-bd.le.edunextstage.net/courses/course-v1:buX+ARC111+2020_Fall/course/

53 Students

Without this PR :
![Screenshot from 2021-12-08 17-13-43](https://user-images.githubusercontent.com/35668326/145303270-46f4418b-c204-4acc-9e0f-4e8437b828f5.png)
With PR:
![Screenshot from 2021-12-08 18-13-29](https://user-images.githubusercontent.com/35668326/145303388-8fddd873-62e4-4dcc-ac3c-1f3e856be696.png)

Without pr 4560 ms and with pr 4174
Generates an improvement of 8.4%

- https://bux-staging-bracu-ac-bd.le.edunextstage.net/courses/course-v1:buX+ARC101+2021_Spring/instructor#view-course_info

21 Students

Without PR:
![Screenshot from 2021-12-08 17-27-37](https://user-images.githubusercontent.com/35668326/145303911-d0d22651-a531-48d3-865e-08cfe77d23c7.png)


With PR:
![Screenshot from 2021-12-08 18-19-20](https://user-images.githubusercontent.com/35668326/145303972-84825997-8f4c-44d5-b4eb-fdcf1561306c.png)

Without pr 4607 ms and with pr 3570
Generates an improvement of 22.5%

- https://openedx-edunext-co.le.edunextstage.net/courses/course-v1:SEGES+AARS2.0+2016_1/instructor#view-course_info

462 Students

Without PR:
![Screenshot from 2021-12-08 17-29-16](https://user-images.githubusercontent.com/35668326/145304265-25a69250-432e-45e4-91b7-72816adc354d.png)

With PR:
![Screenshot from 2021-12-08 18-21-10](https://user-images.githubusercontent.com/35668326/145304319-5601ee28-02a5-4ea9-8d63-4d06d22f1cae.png)

Without pr 1840 ms and with pr: 1385 ms
Generates an improvement of 24.7%