@MaferMazu 2021-12-09
Courses with more enrollments (without pr):

- https://scientiagenus-buap-mx.le.edunextstage.net/courses/course-v1:buap+C101+2020_S2/instructor 
57253 Enrollments but in stage lilac have 38

![Screenshot from 2021-12-09 13-13-36](https://user-images.githubusercontent.com/35668326/145444555-a6c9d411-306d-4d71-b189-6c72f849a16f.png)

- https://lms-academia-pt.le.edunextstage.net/courses/course-v1:academiadigital+EV+acao1/instructor

13691 enrollments

![Screenshot from 2021-12-09 13-26-08](https://user-images.githubusercontent.com/35668326/145445790-03b69223-b655-41ce-b314-4803e7296165.png)

- https://eacademy-spea-ae.le.edunextstage.net/courses/course-v1:SharjahVirtualAcademy+edu124+2020/instructor#view-course_info

13094 enrollments

![Screenshot from 2021-12-09 13-32-57](https://user-images.githubusercontent.com/35668326/145446757-0b0076fb-606b-4b79-aabc-cf8fd42beaa6.png)

## Conclusion

All queries take seconds maybe it is better not to merge this, the change is not very big.