@MaferMazu 2021-12-09
I did not try it with the pr because they told me in the daily that if without the pr it took less than a minute it was not worth keeping ju-9, although I can try with the pr if necessary but I don't think it is more than 30% of improvement, it must be taken into account that everything is in ms, so if something takes 4sec the improvement would be that it last about 3sec.

cc @mariajgrimaldi