@MoisesGSalas 2021-12-09
I'm in favor of closing this PR. The improvements that come from upstream seem good enough.